Eureka !!!! Fix the random results in the inference model of RNN network.
We must keep the states of all layers in the decoder model and to re-inject them in
the t+1 timestamp for every layer. Currently we were setting the last layer state as initial state in t+1,
which is wrong for multilayered architecture. We should set the t+1 state for every layer separately.
Additionally mask layer should be present in the inference model as well.